+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/woowa-techcamp-2020/market-9-blair-dnacu#readme",
   "dependencies": {
+    "cookie-parser": "^1.4.5",
     "express": "^4.17.1",
     "pug": "^3.0.0",
     "serve-favicon": "^2.5.0"
