+++ routes/index.js
@@ -1,0 +1,13 @@
+const mainRouter = require("./main.js");
+const signUpRouter = require("./signUp.js");
+const signInRouter = require("./signIn.js");
+const completeRouter = require("./complete.js");
+const notFoundRouter = require("./notFound.js");
+
+module.exports = {
+  mainRouter,
+  signUpRouter,
+  signInRouter,
+  completeRouter,
+  notFoundRouter,
+};
